+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+* Don't refresh the token if request path is in `accessTokenPaths`.
+* Give `accessTokenKey` and `refreshTokenKey` params value directly in constructor
+* Rename `_accessTokenRoutes` to `_accessTokenPaths` to match public param name.
+
 ## 0.0.1
 
 * Initial release, not tested yet! :)